fix: lifetime contracts (#62)
* docs: drop B4, the transitive-include breakage it claimed does not happen

A facade includes its header in the global module fragment, so those
declarations stay reachable to an importer. Negative control on GCC 14.2:
tests/test_modules.cpp builds with no #include <string> and import rpp.strview.

The 1603-error incident was an import inside a header, which is a different
failure and CLAUDE.md now forbids it. MODULES_MIGRATION.md row 4 carried the
same wrong reasoning and now carries the disproof.

* feature: mark the event_loop borrowed pointers RPP_LIFETIMEBOUND

The loop stores a thread_pool* and an AtomicTimeSource* for its whole life, and
the header stated no rule about either. The constructor and set_time_source()
now carry the attribute, and the doxygen says the loop must not outlive them.

B10 stays open. Measured on clang-18: the attribute warns when a reference
parameter binds a temporary, and it stays silent when a pointer parameter takes
a local that a longer-lived object stores, which is the shape that caused B10.

* docs: record the owner decision that closes B10

A reference overload would let the compiler check the borrow, but the
constructor carries two optional pointers, so that needs one overload per
combination. The documented contract plus the annotation is the practical
limit, and no further work is planned.

* fix: drop lifetimebound from set_time_source, clang rejects it on a void function

clang-20 and later make it a hard error: the attribute needs a return value or a
constructed object to bind to. clang-18 accepts it with no diagnostic, which is
why the local check passed. The constructor keeps the attribute, and the doxygen
still carries the rule.

* revert: drop the pool drain from run_test_func, it cannot work

active_tasks() counts busy WORKERS, not pending WORK: it locks TasksMutex and
counts task->running() over Workers. A job that parallel_task_detached posted
but no worker picked up yet reads 0, so the drain returned at once and left the
race it was added for. The pool exposes no queue depth, so no correct drain can
be written from outside it.

B9 carries the analysis and stays open. Closing it needs a queued work count on
thread_pool, or the join async_task still does not offer.

* fix: race condition in concurrent queue test

* fix: race conditions and project skill setup

* fix: cleanup inefficient buffer use which caused warnings

* fix: address Codex review feedback

* fix: CI failures and intellisense config

* fix: revert the mess caused by codex

* fix: deadlock due to double set_value()

* fix: TSAN false positive

+++ tests/test_file_io.cpp
@@ -717,8 +717,7 @@ TestImpl(test_file_io)
         const int FILE_SIZE = 4096;
         const int TRUNCATE_SIZE = FILE_SIZE / 2;
         std::vector<char> testData = generate_test_data(FILE_SIZE);
-        AssertThat(file::write_new(TestFile, testData.data(), testData.size()),
-                   FILE_SIZE);
+        AssertThat(file::write_new(TestFile, testData.data(), (int)testData.size()), FILE_SIZE);
 
         file f { TestFile, file::READWRITE };
         f.truncate_front(TRUNCATE_SIZE);
@@ -738,8 +737,7 @@ TestImpl(test_file_io)
         const int FILE_SIZE = 4 * 1024 * 1024;
         const int TRUNCATE_SIZE = FILE_SIZE / 2;
         std::vector<char> testData = generate_test_data(FILE_SIZE);
-        AssertThat(file::write_new(TestFile, testData.data(), testData.size()),
-                   FILE_SIZE);
+        AssertThat(file::write_new(TestFile, testData.data(), (int)testData.size()), FILE_SIZE);
 
         file f = { TestFile, file::READWRITE };
         f.truncate_front_sb(TRUNCATE_SIZE);
